requirements.txt
Django==2.2.1
djangorestframework==3.10.3
gunicorn==20.1.0
celery==5.1
django-filter==2.4.0
google-api-python-client==2.8.0
google-api-python-client==2.8.0
google-auth-oauthlib==0.4.4
google-auth-httplib2==0.1.0
django-celery-beat==2.2.0